Move haskell syntax classifier into own package
